view directory splitted
