+++ ssulocker/locker/views/__init__.py
@@ -1,0 +1,3 @@
+from base import *
+from lockerlist import *
+from utils import *
